refactor: CardHolder card -> gamer package로 이동
